Use new banner that better reflects the project
The previous banner incorrectly used the "AACS flag", but that's for Blu-ray Discs, not DVD. No idea why I used it at first, but it is what it is. This new one is more to the point: Disc, Project name, What it is, and What its for. It uses an orange color somewhat akin to VideoLan and VLC. It has also been compressed with almost no perceptual difference so that it can load faster.