Remove set_verbosity_level(), set_cracking_mode(), and related code
It is not actually possible to set the environment variable from python in such a way for the library to get the changes. I've tried manipulating os.environ, setx, py-setenv, and pycrosskit.

Nothing set it persistently for the current terminal/shell in such a way that it could be changed at a whim by these methods. The closest result was setx but it required the entire terminal/shell to reload, making it more or less useless in this context.